+++ 0.56.b/sphere/containers.h
@@ -50,7 +50,7 @@ public:
 		if (empty())
 			return 0;
 
-		int toSkip = 1;
+		size_t toSkip = 1;
 		for (const_iterator it = m_list.begin(); it != m_head && it != m_list.end(); ++it)
 		{
 			if (it == m_list.end())
